callguard-sentinel/backend/main.py
"""
CallGuard Sentinel - Main Backend Service
Orchestrates all microservices and provides the main API gateway
"""

from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import uvicorn
import os
from dotenv import load_dotenv
from fastapi import UploadFile, File


# Load environment variables
load_dotenv()

# Import service modules
from services.audio_ingestion import audio_router
from services.transcription import transcription_router
from services.feature_extraction import feature_router
from services.analysis import analysis_router
from services.alerting import alerting_router
from services.database import init_database

app = FastAPI(
    title="CallGuard Sentinel API",
    description="Real-time scam detection and call analysis system",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure appropriately for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Include service routers
app.include_router(audio_router, prefix="/api/audio", tags=["audio"])
app.include_router(transcription_router, prefix="/api/transcription", tags=["transcription"])
app.include_router(feature_router, prefix="/api/features", tags=["features"])
app.include_router(analysis_router, prefix="/api/analysis", tags=["analysis"])
app.include_router(alerting_router, prefix="/api/alerts", tags=["alerts"])

@app.on_event("startup")
async def startup_event():
    """Initialize database and services on startup"""
    try:
        await init_database()
    except Exception as e:
        print(f"Database not available ({e}). Running without DB - /api/audio/start-call and /analyze/fast/ will still work.")
    print("CallGuard Sentinel backend started successfully")

@app.get("/")
async def root():
    return {"message": "CallGuard Sentinel API", "status": "running"}

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "services": {
            "audio_ingestion": "running",
            "transcription": "running", 
            "feature_extraction": "running",
            "analysis": "running",
            "alerting": "running"
        }
    }

@app.post("/analyze/")
async def analyze_call(file: UploadFile = File(...)):
    content = await file.read()

    if not content:
        return JSONResponse(
            status_code=400,
            content={
                "status": "failed",
                "message": "Uploaded file is empty",
                "success": False,
            },
        )

    result = {
        "status": "completed",
        "message": "Analysis completed successfully",
        "success": True,

        "is_scam": False,
        "final_score": 0.42,
        "alert_level": "warning",
        "confidence": 0.42,
        "analysis_summary": "Demo audio analyzed successfully.",

        "scam_score": 0.42,
        "risk_score": 0.42,
        "risk_level": "warning",
        "prediction": "warning",
        "label": "warning",
        "summary": "Demo audio analyzed successfully.",
        "detected_threats": [
            "Authority indicators detected",
            "Urgency patterns identified",
        ],
        "phone_number": "unknown",
        "duration": 15,
    }

    return JSONResponse(status_code=200, content=result)


@app.post("/analyze/fast/")
async def analyze_call_fast(file: UploadFile = File(...)):
    """Fast analysis endpoint used by CallScreen after a call ends."""
    content = await file.read()

    if not content:
        return JSONResponse(
            status_code=400,
            content={
                "status": "failed",
                "message": "Uploaded file is empty",
                "success": False,
            },
        )

    # Same demo result shape; include fields CallScreen expects: is_fraud, fraud_score, explanation
    result = {
        "status": "completed",
        "success": True,
        "is_scam": False,
        "is_fraud": False,
        "final_score": 0.42,
        "fraud_score": 0.42,
        "scam_score": 0.42,
        "alert_level": "warning",
        "confidence": 0.42,
        "explanation": "Demo audio analyzed successfully. No strong fraud indicators detected.",
        "analysis_summary": "Demo audio analyzed successfully.",
        "summary": "Demo audio analyzed successfully.",
        "risk_score": 0.42,
        "risk_level": "warning",
        "prediction": "warning",
        "label": "warning",
        "detected_threats": [
            "Authority indicators detected",
            "Urgency patterns identified",
        ],
        "phone_number": "unknown",
        "duration": 15,
    }
    return JSONResponse(status_code=200, content=result)


# @app.post("/analyze/")
# async def analyze_call(file: UploadFile = File(...)):
#     content = await file.read()

#     if not content:
#         return {
#             "status": "failed",
#             "message": "Uploaded file is empty",
#         }

#     return {
#         "status": "completed",
#         "is_scam": False,
#         "final_score": 0.42,
#         "alert_level": "warning",
#         "confidence": 0.42,
#         "analysis_summary": "Demo audio analyzed successfully.",
#     }

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True
    )
