callguard-sentinel/backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# websockets==12.0
# python-socketio==5.10.0
# python-multipart==0.0.6
# pydantic==2.5.0
# sqlalchemy==2.0.23
# alembic==1.13.1
# psycopg2-binary==2.9.9
# redis==5.0.1
# celery==5.3.4
# openai-whisper==20231117
# torch==2.1.1
# torchaudio==2.1.1
# transformers==4.36.0
# scikit-learn==1.3.2
# numpy==1.24.4
# pandas==2.1.4
# librosa==0.10.1
# soundfile==0.12.1
# pyaudio==0.2.11
# opensmile==2.3.0
# pyaudioanalyser==0.3.9
# twilio==8.10.0
# python-dotenv==1.0.0
# httpx==0.25.2
# aiofiles==23.2.1

fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-socketio==5.11.4
python-multipart==0.0.9
pydantic==2.9.2
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
nltk==3.9.1
twilio==9.10.2
openai-whisper==20250625
torch==2.10.0
torchaudio==2.10.0
transformers==5.3.0
librosa==0.11.0
soundfile==0.13.1
pyaudio==0.2.14
opensmile==2.6.0
pyaudioanalyser==0.3.9